pima-mnist-mvp-main/requirements.txt
alabaster==0.7.12
attrs==22.1.0
Babel==2.10.3
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
docker-pycreds==0.4.0
docutils==0.17.1
et-xmlfile==1.1.0
fonttools==4.37.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.4
imageio==2.21.3
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==1.1.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.5.3
networkx==2.6.3
numba==0.56.2
numpy==1.21.6
numpy-groupies==0.9.19
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pathtools==0.1.2
Pillow==9.2.0
pluggy==1.0.0
promise==2.3
protobuf==3.20.1
psutil==5.9.2
py==1.11.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.0
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.0
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tomli==2.0.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
typing==3.7.4.3
typing_extensions==4.3.0
urllib3==1.26.12
wandb==0.13.3
zipp==3.8.1
